Some possible changes to the script
I renamed `teams` to `groups` because it was overloading in my head
with Asana teams but doesn't actually have any correlation.

I also got confused on the raw input for the group. I kept putting in
the index number (0, 1) until I actually read the code and saw it wanted
the team name. I added to the helper but not sure if it clarified it.

Added a `quit` so you know something didn't work well if it didn't find
a group.

I haven't written much python but I'd like to make the groups list to
represent the data in oneononeprojects.py and not be hardcoded. That way
it wouldn't be hardcoded in the main file. Thoughts?